Add Http Respone status in Contorller

+++ src/main/java/com/Ratan/my_fist_Demo/Contoller/UsersController.java
@@ -3,7 +3,9 @@ package com.Ratan.my_fist_Demo.Contoller;
 import com.Ratan.my_fist_Demo.DTO.SigupDto;
 import com.Ratan.my_fist_Demo.DTO.UserDto;
 import com.Ratan.my_fist_Demo.Service.UserService;
+import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,31 +19,53 @@ public class UsersController {
     private UserService userService;
 
     @GetMapping
-    public List<UserDto> getAllUsers(){
-        return userService.getAllUsers();
+    public ResponseEntity<?> getAllUsers(){
+        List<UserDto> all=userService.getAllUsers();
+        if(all !=null && !all.isEmpty()){
+            return new ResponseEntity<>(all, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping
-    public UserDto createUser(@RequestBody SigupDto newUser){
-        return userService.createUser(newUser);
+    public ResponseEntity<?> createUser(@RequestBody SigupDto newUser){
+        UserDto user = userService.createUser(newUser);
+        return new ResponseEntity<>(user,HttpStatus.CREATED);
+
     }
 
     @GetMapping("/{id}")
-    public UserDto getUsersBy(@PathVariable Long id){
-        return userService.getUserById(id);
+    public ResponseEntity<?> getUsersBy(@PathVariable Long id){
+        UserDto userById = userService.getUserById(id);
+        if(userById !=null){
+        return new ResponseEntity<>(userById,HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
     }
 
     @PutMapping("/{id}")
-    public UserDto updateUserById(@PathVariable Long id, @RequestBody SigupDto sigupDto){
-        return userService.updateUserbyId(id,sigupDto);
+    public ResponseEntity<?> updateUserById(@PathVariable Long id, @RequestBody SigupDto sigupDto){
+        UserDto userDto = userService.updateUserbyId(id, sigupDto);
+        if(userDto.getId() !=null){
+            return new ResponseEntity<>(userDto,HttpStatus.OK);
+        }
+        else{
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id){
-        userService.deleteByUserId(id);
-        return ResponseEntity.ok("OK");
-
-
+    public ResponseEntity<?> deleteUser(@PathVariable Long id){
+        UserDto userById = userService.getUserById(id);
+        if(userById != null){
+            userService.deleteByUserId(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
 }
